Overhaul routing.
Classes were totally unnecessary and obscured wtf was happening.
The new setup is a condensed form of command/parameters/parser
where multiple commands can have the same parameters/parser,
parameters is the set of values that will be produced by the
parser, and parser is either an inline lambda (for trivial cmds)
or a function defined above that performs more complex transforms.

Because PartialBind always populates every field with None, it's
not necessary for a parser to set a value for every parameter it
provides.

Parameters are primarily used to validate a function against its
desired command - a function that expects a parameter not exposed
by a command cannot bind correctly.