Overhaul Connection.loop
Instead of assuming immediate reconnect is desired, leave it up
to the user.  If the msg comes back empty, trigger a disconnect.
Alternatively, if the message is processed (such as an ERROR) and
we're no longer connected, trigger a disconnect.

This allows users to respond to a disconnect with Client.connect()
before the loop checks connection status to continue.