Make loop required for Connection, EventsMixin
Part of what made tracking down the loop errors in tests was the
fallback to using the default event loop, and then waiting for the
non-default event loop to finish (which will always hang).  Instead,
the base classes Connection and EventsMixin will now require an explicit
loop parameter, to reduce the chance of an inconsistency across a
Client's dependencies.