[impl] slightly smarter parsing for packing templates #74
this takes advantage of `string.Formatter` to parse a string, recovering its literal text, each placeholder name, and
any formatter names.  then, we define some custom formatters to annotate those placeholders with things like:

* `opt` -- optional field, if not provided will be ""
* `join` -- join an iterable into a string with "".join() but leave a bare str alone
* `comma` -- same as join but with "," as a delim
* `space` -- same as join but with " " as a delim
* `bool` -- when value is Truthy, emits the placeholder's name.  when Falsey, emits ""

from that broken down format string, we get a list of literals and formatters.  for example, the template::

    "Hello, {names:comma}!  maybe a: [{a:opt|bool}] {regular} Goodbye!"

will become the following list for rendering::

    [
        "Hello, ",
        [comma_fn],
        "!  maybe a: [",
        [bool_fn],
        "] ",
        [str],
        [" Goodbye!"]
    ]

and we'll also know that `["names", "regular"]` are required, while `["a"]` is optional.

with that information, we have enough to construct parameters and a full command renderer, which should run faster
than the default string formatting (3 functions, 4 str copies, 1 "".join).

to maintain some sanity we don't allow conversions (eg. `"{foo!r}") or nested formatters (eg. `"{foo:{x}}`).

the IRC command structures are not consistent at all, so we have a `deps` argument that can provide an explicit param
dependency graph if the string parsing alone isn't enough.  for example, the WHO command takes a mask and optionally
the flag "o":

    WHO [<mask> [o]]

to represent this in the parser, so that we don't allow sending "WHO o" (which would omit the necessary mask):

    register_serializer_pattern("WHO", "WHO {mask:opt} {o:opt|bool}", deps={"o": "mask"})

some patterns still can't be represented with just dependencies and optional annotations, because extra spaces would
cause problems.  in those cases, we'll register two patterns and score them against the provided kwargs at runtime.
this mainly happens when the command allows an optional argument _before_ a required argument:

    WHOIS [<target>] <mask>
    register_serializer_pattern("WHOIS", "WHOIS {target} {mask:comma}")
    register_serializer_pattern("WHOIS", "WHOIS {mask:comma}")

and there are functions that have different meaning for empty string vs omitted argument::

    AWAY [:[<message>]]
    register_serializer_pattern("AWAY", "AWAY :{message:opt}")
    register_serializer_pattern("AWAY", "AWAY")

where `"AWAY :"` sets the user away with no message, and `"AWAY"` disables the away message.

---

the Serializer can still be cleaned up, but so far the new syntax for command definition is very smooth.  this will
streamline PRs for new commands, and even allowing users to add custom commands from the top level client.